--- src/main/java/ru/aaxee/spring/homework1/domain/service/QuizService.java
@@ -1,11 +1,0 @@
-package ru.aaxee.spring.homework1.domain.service;
-
-import ru.aaxee.spring.homework1.domain.entity.QuizQuestion;
-import ru.aaxee.spring.homework1.domain.entity.QuizResult;
-
-import java.util.List;
-
-public interface QuizService {
-
-    QuizResult run(List<QuizQuestion> quizQuestionList, int maxQuestions);
-}
