--- src/main/java/ru/aaxee/spring/homework1/domain/service/StudentService.java
@@ -1,8 +1,0 @@
-package ru.aaxee.spring.homework1.domain.service;
-
-import ru.aaxee.spring.homework1.domain.entity.Student;
-
-public interface StudentService {
-
-    Student getStudent();
-}
