+++ src/main/java/ru/aaxee/spring/homework1/exception/QuizException.java
@@ -1,4 +1,4 @@
-package ru.aaxee.spring.homework1.infrastructure.exception;
+package ru.aaxee.spring.homework1.exception;
 
 public class QuizException extends Exception {
