+++ src/main/java/ru/aaxee/spring/homework1/service/ConsoleInOutService.java
@@ -1,0 +1,18 @@
+package ru.aaxee.spring.homework1.service;
+
+import java.util.Scanner;
+
+public class ConsoleInOutService implements InOutService {
+
+    private static Scanner scanner = new Scanner(System.in);
+
+    @Override
+    public void write(String text) {
+        System.out.println(text);
+    }
+
+    @Override
+    public String read() {
+        return scanner.nextLine();
+    }
+}
