+++ src/main/java/ru/aaxee/spring/homework1/service/InOutService.java
@@ -1,0 +1,8 @@
+package ru.aaxee.spring.homework1.service;
+
+public interface InOutService {
+
+    void write(String text);
+
+    String read();
+}
