+++ src/main/java/ru/aaxee/spring/homework1/service/StudentService.java
@@ -1,0 +1,8 @@
+package ru.aaxee.spring.homework1.service;
+
+import ru.aaxee.spring.homework1.entity.Student;
+
+public interface StudentService {
+
+    Student getStudent();
+}
