+++ src/main/resources/spring-context.xml
@@ -1,24 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:util="http://www.springframework.org/schema/util"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd">
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <util:constant id="in" static-field="java.lang.System.in"/>
-    <util:constant id="out" static-field="java.lang.System.out"/>
+    <bean id="inOutService" class="ru.aaxee.spring.homework1.service.ConsoleInOutService"/>
 
-    <bean id="studentService" class="ru.aaxee.spring.homework1.domain.service.ConsoleStudentService">
-        <constructor-arg name="inputStream" ref="in"/>
+    <bean id="studentService" class="ru.aaxee.spring.homework1.service.ConsoleStudentService">
+        <constructor-arg name="inOutService" ref="inOutService"/>
     </bean>
 
-    <bean id="quizLoaderService" class="ru.aaxee.spring.homework1.domain.service.CsvQuizLoaderService"/>
+    <bean id="quizLoaderService" class="ru.aaxee.spring.homework1.service.CsvQuizLoaderService"/>
 
-    <bean id="quizService" class="ru.aaxee.spring.homework1.domain.service.ConsoleQuizService">
-        <constructor-arg name="inputStream" ref="in"/>
-        <constructor-arg name="printStream" ref="out"/>
+    <bean id="quizService" class="ru.aaxee.spring.homework1.service.ConsoleQuizService">
+        <constructor-arg name="inOutService" ref="inOutService"/>
     </bean>
 
-    <bean id="resultPrintService" class="ru.aaxee.spring.homework1.domain.service.ConsoleResultPrintService">
+    <bean id="resultPrintService" class="ru.aaxee.spring.homework1.service.ConsoleResultPrintService">
         <constructor-arg name="congratulation" value="Поздравляем!"/>
-        <constructor-arg name="printStream" ref="out"/>
+        <constructor-arg name="inOutService" ref="inOutService"/>
     </bean>
 </beans>
